Fix bug on getHash() function on HashingUtils
When the input had a length that is multiple of the buffer size,
it would throw a indexOutOfBoundException.

The problem was that the read() function returned -1 if the inputReader
was at the EOF and it wasn't checked, when calling the update method from
MessageDigest with the -1 as the len param, it throwed the exception.

The solution was checking for EOF instead of checking that readBytes
was different than the buffer size, and changing the do..while to a while loop